moved delayLog to view folder
