+++ blitz-build/src/components/weather/DashboardWeather.jsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import reverseGeocode from "reverse-geocode";
 
 // for dashboard import <Weather usage="dashboard"/>
-// for project page import <Weather usage="project" city={} latitude={} longitude={} />
+
 
 function DashboardWeather({ weatherData, time, icon, weatherPosition }) {
   const cityInfo = reverseGeocode.lookup(
