+++ blitz-build/src/views/delayLog/DelayLog.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useContext } from "react";
 
 //components
-import DelayLogButton from "./DelayLogButton";
+import DelayLogButton from "../../components/delayLog/DelayLogButton";
 
 //contexts
 import DelayLogContext from "../../contexts/delayLog/DelayLogContext";
@@ -21,7 +21,7 @@ import TableFooter from "@material-ui/core/TableFooter";
 import TablePagination from "@material-ui/core/TablePagination";
 
 // pages bar function from global
-import TablePaginationActions from "../global/TablePaginationActions";
+import TablePaginationActions from "../../components/global/TablePaginationActions";
 // csv function
 import { ExportToCsv } from "export-to-csv";
